+++ src/whispercpp/__init__.py
@@ -59,5 +59,13 @@ class Whisper:
     def transcribe_from_file(self, filename: str, num_proc: int = 1):
         return self.transcribe(api.load_wav_file(filename).mono, num_proc)
 
+    async def capture_audio(
+        self, device_id: int | None = None, sample_rate: int | None = None
+    ):
+        if device_id is None:
+            device_id = -1
+        if sample_rate is None:
+            sample_rate = api.SAMPLE_RATE
+
 
 __all__ = ["Whisper", "api"]
