+++ src/whispercpp/__init__.pyi
@@ -16,5 +16,13 @@ class Whisper:
     def transcribe_from_file(self, filename: str) -> str: ...
     @overload
     def transcribe_from_file(self, filename: str, num_proc: int = ...) -> str: ...
+    @overload
+    async def capture_audio(self) -> None: ...
+    @overload
+    async def capture_audio(self, device_id: int | None = ...) -> None: ...
+    @overload
+    async def capture_audio(
+        self, device_id: int | None = ..., sample_rate: int | None = ...
+    ) -> None: ...
     @classmethod
     def from_pretrained(cls, model_name: str) -> Whisper: ...
